% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/read_bankscope.R
\name{read_bankscope}
\alias{read_bankscope}
\title{Reads an Excel-export from Bankscope}
\usage{
read_bankscope(path, escape_vars = "Shareholder")

read_bankscope(path, escape_vars = "Shareholder")
}
\arguments{
\item{path}{the path to a file}

\item{escape_vars}{a vector of column (variable) names that have a \% or a currency key but that should not be treated as a variable, defaults to "Shareholder"}

\item{path}{the path to a file}

\item{escape_vars}{a vector of column (variable) names that have a \% or a currency key but that should not be treated as a variable, defaults to "Shareholder"}
}
\value{
a data.table in long-format

a data.table in long-format
}
\description{
Be aware that the file has to be manually opened and saved first (Dunno why...)

Be aware that the file has to be manually opened and saved first (Dunno why...)
}
\examples{
read_bankscope("Bankscope_Export_1.xslx")
read_bankscope("Bankscope_Export_1.xslx", escape_vars = c("Shareholders", "Number of Branches"))

read_bankscope("Bankscope_Export_1.xslx")
read_bankscope("Bankscope_Export_1.xslx", escape_vars = c("Shareholders", "Number of Branches"))

}
\seealso{
to read all files within a folder, use \code{\link{read_folder}}

to read all files within a folder, use \code{\link{read_folder}}
}

