% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_folder.R
\name{read_folder}
\alias{read_folder}
\title{Read all files in a folder}
\usage{
read_folder(path_folder, ..., ncores = NA)
}
\arguments{
\item{path_folder}{the path to a folder}

\item{...}{additional parameters that are passed to \code{\link{read_bankscope}}}

\item{ncores}{number of cores that are used for reading the data-files, defaults to the number of cores as returned by \code{parallel::detectCores(logical = TRUE)}}
}
\value{
a data.table in long-format
}
\description{
Be aware that all files have to be of type .xlsx and have to be manually opened and saved first (dunno why)
}
\examples{

read_folder("data")

read_folder("data/")

read_folder("data", ncores = 2)

read_folder("data", escape_vars = c("Shareholder", "Number of Branches"))
}

