% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_data.R
\name{save_data}
\alias{save_data}
\title{Saves the (raw)-data from Bankscope into a csv or zip-file}
\usage{
save_data(dt, file, folder = NA)
}
\arguments{
\item{dt}{a data.table as outputted from read_folder or read_bankscope}

\item{file}{the name to a file (with file-extension (.zip, .RDS, .csv))}

\item{folder}{the path to the folder, defaults to current working directory}
}
\value{
nothing
}
\description{
Saves the (raw)-data from Bankscope into a csv or zip-file
}
\examples{
dt <- read_folder("folder_to_my_exports")
save_data(dt, file = "my_data.zip")
save_data(dt, file = "my_data.RDS")
save_data(dt, file = "my_data.csv")
}
\seealso{
all file-formats can be read again with \code{\link{load_data}}
}

