% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_files.R
\name{zip_files}
\alias{zip_files}
\title{Compresses files to a .zip-files in the current folder}
\usage{
zip_files(zip_name, ..., file_names = NA, folder = NA)
}
\arguments{
\item{zip_name}{the output name for the .zip-file}

\item{...}{data.frames that can be written to csv using fwrite}

\item{file_names}{a vector of names for the .csv-files
(if left empty, the files will be named file1.csv, file2.csv etc.)}

\item{folder}{the folder, where the file will be saved
#'}
}
\value{
nothing
}
\description{
Compresses files to a .zip-files in the current folder
}
\examples{
x <- data.frame(x = 1:10)
y <- data.frame(y = 1:10)

zip_files(zip_name = "my_zip", x, y)
}

